Move pylxd.hosts to pylxd.deprecated.hosts
