+++ pylxd/deprecated/hosts.py
